Consolidate group quiz question and quiz question report components. Add missing resource handling.
